PO-1993 remove unrelated opal minor creditor changes
